Add new fancy printing bash lib
Source lib/fancy.sh to get access to fancy_print and fancy_println functions

Usage: fancy_print <style> <color> <value>

At least 1 parameter is required and that is <value>
If there are 2 arguments, then the 1st argument will be used as a style. If the first
argument is not a style, then it will be reattempted as a color.

Examples:
 - Full Usage: fancy_print "bold" "red" "hello world"
 - 2 Arg Usage: fancy_print "underline" "hello world"
 - 2 Arg Usage: fancy_print "yellow" "hello world"
 - 1 Arg Usage: fancy_print "hello world"

Note: fancy_print will not add a newline. To add a newline, use fancy_println.